requirements.txt
# OpenBlog Neo - Dependencies

# Core
pydantic>=2.0
python-dotenv>=1.0

# HTTP
httpx>=0.25

# AI
google-genai>=1.0

# XML parsing (secure)
defusedxml>=0.7

# Export formats
markdownify>=0.11
openpyxl>=3.1

# API Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Legal Engine (for law firm content)
browser-use>=0.11.3
playwright>=1.40.0
langchain-google-genai>=1.0.0

# Google Drive API (for webinar processing)
google-api-python-client>=2.0
google-auth>=2.0
